+++ src/material/dialog/testing/BUILD.bazel
@@ -25,13 +25,12 @@ filegroup(
 
 ng_test_library(
     name = "unit_tests_lib",
-    srcs = glob(
-        ["**/*.spec.ts"],
-        exclude = ["shared.spec.ts"],
-    ),
+    srcs = glob(["**/*.spec.ts"]),
     deps = [
-        ":harness_tests_lib",
         ":testing",
+        "//src/cdk/overlay",
+        "//src/cdk/testing",
+        "//src/cdk/testing/testbed",
         "//src/material/dialog",
         "@npm//@angular/platform-browser",
     ],
@@ -41,18 +40,5 @@ ng_web_test_suite(
     name = "unit_tests",
     deps = [
         ":unit_tests_lib",
-    ],
-)
-
-ng_test_library(
-    name = "harness_tests_lib",
-    srcs = ["shared.spec.ts"],
-    deps = [
-        ":testing",
-        "//src/cdk/overlay",
-        "//src/cdk/testing",
-        "//src/cdk/testing/testbed",
-        "//src/material/dialog",
-        "@npm//@angular/platform-browser",
     ],
 )
